+++ pkg/mock/IcecreamHasIngredientsServiceMock.go
@@ -1,11 +1,11 @@
 package mock
 
 type IcecreamHasIngredientsService struct {
-	CreateFn      func(icecreamProductId int, ingredientIds []int) error
+	CreateFn      func(icecreamProductId int64, ingredientIds []int64) error
 	CreateInvoked bool
 }
 
-func (s *IcecreamHasIngredientsService) Create(icecreamProductId int, ingredientIds []int) error {
+func (s *IcecreamHasIngredientsService) Create(icecreamProductId int64, ingredientIds []int64) error {
 	s.CreateInvoked = true
 	return s.CreateFn(icecreamProductId, ingredientIds)
 }
