+++ pkg/mock/SourcingValuesServiceMock.go
@@ -3,33 +3,33 @@ package mock
 import "github.com/fraenky8/zlr-ca/pkg/domain"
 
 type SourcingValueService struct {
-	CreatesFn      func(sourcingValues domain.SourcingValues) ([]int, error)
+	CreatesFn      func(sourcingValues domain.SourcingValues) ([]int64, error)
 	CreatesInvoked bool
 
-	ReadFn      func(icecreamProductId int) (domain.SourcingValues, error)
+	ReadFn      func(icecreamProductId int64) (domain.SourcingValues, error)
 	ReadInvoked bool
 
-	ReadsFn      func(icecreamProductIds []int) ([]domain.SourcingValues, error)
+	ReadsFn      func(icecreamProductIds []int64) ([]domain.SourcingValues, error)
 	ReadsInvoked bool
 
 	ReadAllFn      func() (domain.SourcingValues, error)
 	ReadAllInvoked bool
 
-	DeletesFn      func(icecreamProductIds []int) error
+	DeletesFn      func(icecreamProductIds []int64) error
 	DeletesInvoked bool
 }
 
-func (s *SourcingValueService) Creates(sourcingValues domain.SourcingValues) ([]int, error) {
+func (s *SourcingValueService) Creates(sourcingValues domain.SourcingValues) ([]int64, error) {
 	s.CreatesInvoked = true
 	return s.CreatesFn(sourcingValues)
 }
 
-func (s *SourcingValueService) Reads(icecreamProductIds []int) ([]domain.SourcingValues, error) {
+func (s *SourcingValueService) Reads(icecreamProductIds []int64) ([]domain.SourcingValues, error) {
 	s.ReadsInvoked = true
 	return s.ReadsFn(icecreamProductIds)
 }
 
-func (s *SourcingValueService) Read(icecreamProductIds int) (domain.SourcingValues, error) {
+func (s *SourcingValueService) Read(icecreamProductIds int64) (domain.SourcingValues, error) {
 	s.ReadsInvoked = true
 	return s.ReadFn(icecreamProductIds)
 }
@@ -39,7 +39,7 @@ func (s *SourcingValueService) ReadAll() (domain.SourcingValues, error) {
 	return s.ReadAllFn()
 }
 
-func (s *SourcingValueService) Deletes(icecreamProductIds []int) error {
+func (s *SourcingValueService) Deletes(icecreamProductIds []int64) error {
 	s.DeletesInvoked = true
 	return s.DeletesFn(icecreamProductIds)
 }
